+++ packages/ui-pager/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [14.1.24](https://github.com/nativescript-community/ui-pager/compare/v14.1.23...v14.1.24) (2024-05-14)
+
+**Note:** Version bump only for package @nativescript-community/ui-pager
+
+## [14.1.23](https://github.com/nativescript-community/ui-pager/compare/v14.1.22...v14.1.23) (2024-04-17)
+
+**Note:** Version bump only for package @nativescript-community/ui-pager
+
+## [14.1.22](https://github.com/nativescript-community/ui-pager/compare/v14.1.21...v14.1.22) (2024-03-22)
+
+**Note:** Version bump only for package @nativescript-community/ui-pager
+
+## [14.1.21](https://github.com/nativescript-community/ui-pager/compare/v14.1.20...v14.1.21) (2024-03-22)
+
+**Note:** Version bump only for package @nativescript-community/ui-pager
+
+## [14.1.20](https://github.com/nativescript-community/ui-pager/compare/v14.1.19...v14.1.20) (2024-03-01)
+
+**Note:** Version bump only for package @nativescript-community/ui-pager
+
 ## [14.1.19](https://github.com/nativescript-community/ui-pager/compare/v14.1.18...v14.1.19) (2024-02-15)
 
 **Note:** Version bump only for package @nativescript-community/ui-pager